# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is the API documentation for the kidzzle project.",
        "title": "kidzzle API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/assessments/{assessmentType}/question/": {
            "get": {
                "description": "Get assessment question by assessment type",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Assessment"
                ],
                "summary": "Get assessment question",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Assessment Type",
                        "name": "assessmentType",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Assessment question retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/domain.GetAssessmentQuestionResponse"
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "schema": {
                            "$ref": "#/definitions/domain.ValidationError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized error",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponseExample401"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponseExample500"
                        }
                    }
                }
            }
        },
...
        "/assessments/{kidId}/{assessmentQuestionId}/result": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get an assessment result for a given kid and assessment question",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Assessment"
                ],
                "summary": "Get an assessment result by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Access Token",
                        "name": "accessToken",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Kid ID",
                        "name": "kidId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Assessment Question ID",
                        "name": "assessmentQuestionId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Assessment result retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/domain.GetAssessmentResultByIdResponse"
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "schema": {
                            "$ref": "#/definitions/domain.ValidationError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized error",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponseExample401"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponseExample500"
                        }
                    }
                }
            }
        },
...
        "/kids/{kidId}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update the data of a kid by their ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Kid"
                ],
                "summary": "Update kid data by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Access Token",
                        "name": "accessToken",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Kid ID",
                        "name": "kidId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Kid Data Update Request",
                        "name": "kidData",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.UpdateKidDataByIdRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Kid data updated successfully",
                        "schema": {
                            "$ref": "#/definitions/domain.UpdateKidDataByIdResponse"
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "schema": {
                            "$ref": "#/definitions/domain.ValidationError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized error",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponseExample401"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponseExample500"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete kid data by given kid ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Kid"
                ],
                "summary": "Delete kid data by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Access Token",
                        "name": "accessToken",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Kid ID",
                        "name": "kidId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Kid data deleted successfully",
                        "schema": {
                            "$ref": "#/definitions/domain.DeleteKidDataByIdResponse"
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "schema": {
                            "$ref": "#/definitions/domain.ValidationError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized error",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponseExample401"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponseExample500"
                        }
                    }
                }
            }
        },
        "/mother-pregnant/{pregnantId}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete a mother pregnant record by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MotherPregnant"
                ],
                "summary": "Delete a mother pregnant record by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Access Token",
                        "name": "accessToken",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Pregnant ID",
                        "name": "pregnantId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Mother pregnant record deleted successfully",
                        "schema": {
                            "$ref": "#/definitions/domain.DeleteMotherPregnantByIdResponse"
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "schema": {
                            "$ref": "#/definitions/domain.ValidationError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized error",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponseExample401"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponseExample500"
                        }
                    }
                }
            }
        },
...
        "/mother-pregnants/create": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new mother pregnant record with the given details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MotherPregnant"
                ],
                "summary": "Create a new mother pregnant record",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Access Token",
                        "name": "accessToken",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Mother Pregnant Request",
                        "name": "motherPregnant",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.CreateMotherPregnantRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Mother pregnant record created successfully",
                        "schema": {
                            "$ref": "#/definitions/domain.CreateMotherPregnantResponse"
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "schema": {
                            "$ref": "#/definitions/domain.ValidationError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized error",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponseExample401"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponseExample500"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "domain.CreateMotherPregnantRequest": {
            "type": "object",
            "properties": {
                "accessToken": {
                    "type": "string"
                },
                "mother_birthday": {
                    "type": "string"
                },
                "mother_name": {
                    "type": "string"
                },
                "pregnant_complications": {
                    "type": "string"
                },
                "pregnant_congenital_disease": {
                    "type": "string"
                },
                "pregnant_drug_history": {
                    "type": "string"
                },
                "userId": {
                    "type": "string"
                }
            }
        },
        "domain.CreateMotherPregnantResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "message": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
        "domain.DeleteKidDataByIdResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "message": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
        "domain.DeleteMotherPregnantByIdResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "message": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
...
        "domain.ErrorResponseExample401": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 401
                },
                "message": {
                    "type": "string",
                    "example": "Unauthorized"
                }
            }
        },
        "domain.ErrorResponseExample500": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 500
                },
                "message": {
                    "type": "string",
                    "example": "Internal Server Error"
                }
            }
        },
...
        "domain.GetAssessmentQuestionResponse": {
            "type": "object",
            "properties": {
                "age_range_id": {
                    "type": "string",
                    "example": "AGER_4"
                },
                "age_range_name": {
                    "type": "string",
                    "example": "3-4"
                },
                "assessment_method": {
                    "type": "string",
                    "example": "\u003cp\u003e1. จัดให้เด็กอยู่ในท่านอนคว่ำบนพื้นราบ\u0026nbsp;\u003c/p\u003e\u003cp\u003e2. เขย่ากรุ๋งกริ๋งด้านหน้าเด็กเพื่อให้เด็กสนใจ แล้วเคลื่อนขึ้นด้านบน กระตุ้นให้เด็กมองตาม\u003c/p\u003e"
                },
                "assessment_no": {
                    "type": "string",
                    "example": "1"
                },
                "assessment_question_id": {
                    "type": "string",
                    "example": "088e7d6b-faa8-11ef-870e-0242ac120002"
                },
                "assessment_required_tool": {
                    "type": "string",
                    "example": "กรุ๋งกริ๋ง"
                },
                "assessment_type_id": {
                    "type": "string",
                    "example": "ASSMTT_1"
                },
                "assessment_type_name": {
                    "type": "string",
                    "example": "DAIM"
                },
                "dev_type_id": {
                    "type": "string",
                    "example": "DEVT_1"
                },
                "development_type": {
                    "type": "string",
                    "example": "GM"
                },
                "pass_criteria": {
                    "type": "string",
                    "example": "เด็กยกศีรษะและอกโดยใช้แขนยันกับพื้นพยุงตัวไว้อย่างน้อย 5 วินาที"
                },
                "question_text": {
                    "type": "string",
                    "example": "ท่านอนคว่ำยกศีรษะและอกพ้นพื้น (GM)"
                }
            }
        },
        "domain.GetAssessmentResultByIdResponse": {
            "type": "object",
            "properties": {
                "assessment_question_id": {
                    "type": "string",
                    "example": "3afe2697-2b02-46d4-a9c3-efd0bdd9e014"
                },
                "assessment_result_id": {
                    "type": "string",
                    "example": "8457529e-01fb-41d0-82b1-9618be667862"
                },
                "created_at": {
                    "type": "string",
                    "example": "2025-03-09T00:00:00Z"
                },
                "is_passed": {
                    "type": "boolean",
                    "example": true
                },
                "kid_id": {
                    "type": "string",
                    "example": "7d7bc87f-0faf-43f9-9af9-1a09166db993"
                }
            }
        },
...
        "domain.UpdateKidDataByIdRequest": {
            "type": "object",
            "properties": {
                "accessToken": {
                    "type": "string"
                },
                "kidId": {
                    "type": "string"
                },
                "kid_birth_weight": {
                    "type": "string"
                },
                "kid_birthday": {
                    "type": "string"
                },
                "kid_blood_type": {
                    "type": "string"
                },
                "kid_body_length": {
                    "type": "string"
                },
                "kid_congenital_disease": {
                    "type": "string"
                },
                "kid_gender": {
                    "type": "string"
                },
                "kid_gestational_age": {
                    "type": "string"
                },
                "kid_name": {
                    "type": "string"
                },
                "kid_oxygen": {
                    "type": "string"
                },
                "pregnant_id": {
                    "type": "string"
                },
                "userId": {
                    "type": "string"
                }
            }
        },
        "domain.UpdateKidDataByIdResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "message": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
...
        "domain.ValidationError": {
            "type": "object",
            "properties": {
                "errors": {
                    "$ref": "#/definitions/domain.ValidationErrorResponse"
                }
            }
        },
        "domain.ValidationErrorResponse": {
            "type": "object",
            "properties": {
                "auth_error": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "fullname": {
                    "type": "string"
                },
                "incorrect": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "user_error": {
                    "type": "string"
                }
            }
        }
    }
}